Prevent death by SIGHUP
If a udev rule is configured to send SIGHUP to evmapy when a new input
device is connected and /dev/uinput is writable for the user which
evmapy is run as, evmapy will be killed upon startup because the SIGHUP
handler isn't installed until the run() method call. Therefore we should
simply ignore SIGHUP while initializing the multiplexer.