Action triggering using event sequences
An action can now be triggered when multiple events occur in an
uninterrupted sequence.

Note that the first event in the sequence is never checked immediately -
if the second or further event matches, its predecessor is then checked.
That's why `sequence_cur' is reset to 1 rather than 0. While this
doubles the number of checks performed, it's the cleanest implementation
I could come up with which handles all cases correctly.