Allow partial configuration to be loaded
Requiring a list of all input device axes and buttons to be present in
every configuration file loaded makes little sense as device
capabilities do not change while the program is running. Thus the "axes"
and "buttons" keys are now only required in the initially loaded
configuration file, while the configuration files loaded dynamically are
only required to have the "actions" and "grab" keys set.