Stale control socket removal
When the application is shut down abruptly, the control socket is not
removed from the filesystem. The next time the application is launched
that stale socket will prevent it from running. Thus before binding to
the control socket, we try to send a control request to an already
existing one. If a control socket exists, but sending the control
request fails, it means the socket is stale and can be removed.